Various
Improved web integrated clipboard
Fully fledged brush visualization (modes 0-2): //br vis
Brush targeting modes (0-3): //br target
Brush scroll actions: //br scroll
- change clipboard from directory or web
- change mask
- change pattern
- change range
- change size
- change target mode
Build multi-brushes (choose the brush for left click and right click
individually)
- //br primary
- //br secondary
- Changing the pattern/mask will affect the currently selected brush,
not both
Changes to brush behavior
- spline connects by clicking the same spot twice
- line/copy brush tweaks
Schematics now default per user
- To save/load globally use "../" before the filename
Easily save a schematic to a directory
- `//schem save folder/` will choose save a the next lowest free number
Improved block parsing
Add resettable brushes
- e.g. line brush resets points by shift + left click

+++ core/src/main/java/com/boydti/fawe/object/PseudoRandom.java
@@ -31,7 +31,7 @@ public class PseudoRandom {
     }
 
     public double nextDouble() {
-        return Math.max(0, Math.min(1, Math.abs(nextLong() / Long.MAX_VALUE)));
+        return Math.max(0, Math.min(1, Math.abs((double) nextLong() / Long.MAX_VALUE)));
     }
 
     public int random(final int n) {
